feat: add JSDoc rules for React components and hooks
