+++ lib/index.js
@@ -22,6 +22,8 @@ const plugin = {
 
     //documentation
     "require-jsdoc-on-root-function": require("./rules/documentation/require-jsdoc-on-root-function"),
+    "require-jsdoc-on-component": require("./rules/documentation/require-jsdoc-on-component"),
+    "require-jsdoc-on-hook": require("./rules/documentation/require-jsdoc-on-hook"),
   },
 };
 
@@ -44,6 +46,8 @@ plugin.configs = {
       "eslint-frontend-rules/no-nested-component": "error",
       "eslint-frontend-rules/no-unnecessary-fragment": "warn",
       "eslint-frontend-rules/require-jsdoc-on-root-function": "warn",
+      "eslint-frontend-rules/require-jsdoc-on-component": "warn",
+      "eslint-frontend-rules/require-jsdoc-on-hook": "warn",
     },
   },
 };
