+++ lib/utils/react.js
@@ -86,10 +86,15 @@ function isHookName(name) {
   return name && /^use[A-Z0-9]/.test(name);
 }
 
+function isRootFunction(name) {
+  return name && !isComponentName(name) && !isHookName(name);
+}
+
 module.exports = {
   getMeaningfulChildren,
   isFragmentElement,
   isComponent,
   isComponentName,
+  isRootFunction,
   isHookName,
 };
